% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ghypMarkov}
\alias{ghypMarkov}
\title{ghypMarkov}
\usage{
ghypMarkov(data, response, family, ...)
}
\arguments{
\item{data}{data frame or matrix of multivariate series (eg, stock returns)}

\item{response}{equations for series in depmix() function}

\item{family}{distributions for depmix() function}
}
\value{
list
}
\description{
Fit Hidden Markov Models under symmetric generalized hyperbolic distribution
}
\details{
This function fit Hidden Markov Models under symmetric generalized hyperbolic distribution by first fitting the 2-state Markov model assuming Gaussian distribution, then fitting the two state subsets of data to symmetric generalized hyberbolic distribution separately, finally recomputing posterior probabilities of hidden Markov State
}
\author{
Stephen Downing
}
\keyword{Markov}
\keyword{model}

