% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mapBetween}
\alias{mapBetween}
\title{mapBetween}
\usage{
mapBetween(x, minmax = c(1, 20), log = F)
}
\arguments{
\item{x}{an n-length vector of numeric values}

\item{minmax}{a vector of two positive numeric values defining the interval
onto which x is mapped, e.g. c(1,20)}

\item{log}{logical to do log transform; default FALSE}
}
\value{
a vector same length as x argument
}
\description{
Map a vector (e.g., vertex sizes) proportionally or via log transform
to a given real interval.
}
\details{
This function (f: Rd -> Rd) maps a vector proportionally or via log transform
to a given real interval. Output values of -Inf are coded as the minimum of
the minmax argument. Nonpositive (x_i < 0) elements of input vector are okay
since log transform is done after mapping to minmax interval, which must be
positive. Use log transform to even out the intervals between elements in x,
for example to make more clear sizes of graph vertices.
}
\examples{
\dontrun{
## simple example
x1 <- 90:100
mapBetween(x = x1, minmax = c(1,20), log = F)
## log transform with negative input values
x2 <- -5:5
mapBetween(x = x2, minmax = c(0,1), log = T)
## log transform with negative decimal values
x3 <- seq(from = -1.5, to = 1.5, by = .1)
mapBetween(x = x3, minmax = c(1,2), log = T)
}
}
\author{
Stephen Downing
}

