% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiStockHistory}
\alias{multiStockHistory}
\title{multiStockHistory}
\usage{
multiStockHistory(symbols, symbols.names = NULL, startday = 1,
  startmonth = 1, startyear = 1970, endday = 19, endmonth = 10,
  endyear = 2014, series = "Adj.Close", ...)
}
\arguments{
\item{symbols}{char vector of stock or index symbols}

\item{symbols.names}{char vector of names for symbols}

\item{startday}{int first day of series}

\item{startmonth}{int}

\item{startyear}{int}

\item{endday}{int}

\item{endmonth}{int}

\item{endyear}{int}

\item{series}{char default "Adj.Close" {"High","Low","Close","Volume","Adj.Close"}}
}
\value{
list: [[1]] data.frame of series attribute from all symbols, [[2]] list of data.frames of all series from all symbols
}
\description{
Pull multiple historic stock returns series
}
\details{
This function downloads multiple historic series for stocks and stock indices from Yahoo, returning a returns list: [[1]] data frame of given series from all symbols [[2]] full list of all series data frames from all symbols
}
\examples{
\dontrun{
multiStockHistory(c('AAPL','GOOG'))
tail(y$df)
}
}
\author{
Stephen Downing
}
\keyword{returns}
\keyword{stock}

