\name{sdowninMisc-package}
\alias{sdowninMisc-package}
\alias{sdowninMisc}
\docType{package}
\title{
The Personal R package of Stephen Downing
}
\description{
A personal R package with miscellaneous functions for data wrangling, analysis, and simulation.
}
\details{
\tabular{ll}{
Package: \tab sdowninMisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-19\cr
License: \tab GPL (v3 or later)\cr
}
Miscellaneous functions for data wrangling (e.g., multiStockHistory(), mapBetween()), analysis (e.g., ghypMarkov()), and simulation.
}
\author{
Stephen Downing

Maintainer: Stephen Downing <sdowning.bm02g@nctu.edu.tw>
}
\references{
ghyp, depmixS4
}
Markov models, stock returns
\keyword{ package }
\seealso{
   \code{\link[depmix:dempixS4-package]{depmixS4}}
   \code{\link[ghyp-package]{ghyp}}
}
\examples{
ghypMarkov()
}
