\name{sdowninMisc-package}
\alias{sdowninMisc-package}
\alias{sdowninMisc}
\docType{package}
\title{
What the package does 
}
\description{
More about what it does 
}
\details{
\tabular{ll}{
Package: \tab sdowninMisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-19\cr
License: \tab GPL (v3 or later)\cr
}
overview and important functions
}
\author{
Stephen Downing

Maintainer: Stephen Downing <sdowning.bm02g@nctu.edu.tw>
}
\references{
ghyp, depmixS4
}
Markov models, stock returns
\keyword{ package }
\seealso{
   \code{\link[depmix:dempixS4-package]{depmixS4}}
   \code{\link[ghyp-package]{ghyp}}
}
\examples{
ghypMarkov()
}
